requirements_dev.txt
pip==18.1
bumpversion==0.5.3
wheel==0.32.1
watchdog==0.9.0
flake8==3.5.0
tox==3.5.2
coverage==4.5.1
Sphinx==1.8.1
twine==1.12.1

pytest==3.8.2
pytest-runner==4.2

pyyaml>=4.2b1
tensorflow==1.12.0
tqdm==4.29.1
neo4j-driver==1.6.2
fuzzywuzzy==0.17.0
bert-serving-server==1.7.9
bert-serving-client==1.7.9
